Move CreateTask to is own WorkItemUpdateService
