+++ Satori/Pages/StandUp/Components/EditWorkItem.razor
@@ -5,6 +5,7 @@
 @inject IJSRuntime JsRuntime
 @inject UserService UserService
 @inject StandUpService StandUpService
+@inject WorkItemUpdateService WorkItemUpdateService
 
 @if (ViewModel.WorkItem == null)
 {
