+++ Satori/Utilities/AppServiceBuilder.cs
@@ -17,6 +17,7 @@ internal static class AppServiceBuilder
         services.AddScoped<PullRequestService>();
         services.AddScoped<UserService>();
         services.AddScoped<StandUpService>();
+        services.AddScoped<WorkItemUpdateService>();
         
         services.AddScoped(serviceProvider => serviceProvider.GetRequiredService<IConnectionSettingsStore>().GetAzureDevOpsSettings());
         services.AddScoped<IAzureDevOpsServer, AzureDevOpsServer>();
